Record build size for all commits on master
The executable size for builds on Clang and GCC on Linux, Clang on
macOS, and Emscripten will be recorded in a Google Sheet as part of the
build process on CI.

Unfortunately Travis CI's Windows build integration does not play nicely
with secrets. The build hangs, see:

https://travis-ci.community/t/current-known-issues-please-read-this-before-posting-a-new-topic/264/10

So we can't record the size data on Windows.